some views refactoring
