+++ etf/templates/documents.html
@@ -5,9 +5,10 @@
   <form method="POST">{{csrf_input}}
     <fieldset class="fieldset">
       <legend>
-        <h3>Evaluation costs and budget - PLACEHOLDER</h3>
+        <h3>Documents - PLACEHOLDER</h3>
       </legend>
-
+      To add multiple documents (?)
+      Document upload to follow.
       {{macros.buttons(prev_url)}}
     </fieldset>
   </form>
