+++ etf/templates/publication-intention.html
@@ -5,9 +5,8 @@
   <form method="POST">{{csrf_input}}
     <fieldset class="fieldset">
       <legend>
-        <h3>Evaluation costs and budget - PLACEHOLDER</h3>
+        <h3>Publication intention - PLACEHOLDER</h3>
       </legend>
-
       {{macros.buttons(prev_url)}}
     </fieldset>
   </form>
