+++ etf/urls.py
@@ -1,10 +1,10 @@
 from django.contrib import admin
 from django.urls import include, path
 
-from etf.evaluation import views
+from etf.evaluation import views, submission_views
 
 urlpatterns = [
-    path("", views.index_view, name="index"),
+    path("", submission_views.index_view, name="index"),
     path("admin/", admin.site.urls),
     path("accounts/", include("allauth.urls")),
     path("search/", views.search_evaluations_view, name="search"),
@@ -42,64 +42,81 @@ api_urlpatterns = [
 ]
 
 evaluation_entry_urlpatterns = [
-    path("evaluation/<uuid:evaluation_id>/", views.intro_page_view, name="intro"),
-    path("evaluation/<uuid:evaluation_id>/title", views.evaluation_title_view, name="title"),
-    path("evaluation/<uuid:evaluation_id>/description", views.evaluation_description_view, name="description"),
+    path("evaluation/<uuid:evaluation_id>/", submission_views.intro_page_view, name="intro"),
+    path("evaluation/<uuid:evaluation_id>/title", submission_views.evaluation_title_view, name="title"),
+    path(
+        "evaluation/<uuid:evaluation_id>/description", submission_views.evaluation_description_view, name="description"
+    ),
     path(
         "evaluation/<uuid:evaluation_id>/issue-description",
-        views.evaluation_issue_description_view,
+        submission_views.evaluation_issue_description_view,
         name="issue-description",
     ),
     path(
         "evaluation/<uuid:evaluation_id>/studied-population",
-        views.evaluation_studied_population_view,
+        submission_views.evaluation_studied_population_view,
         name="studied-population",
     ),
     path(
         "evaluation/<uuid:evaluation_id>/participant-recruitment",
-        views.evaluation_participant_recruitment,
+        submission_views.evaluation_participant_recruitment,
         name="participant-recruitment",
     ),
     path(
         "evaluation/<uuid:evaluation_id>/evaluation-costs",
-        views.evaluation_costs_view,
+        submission_views.evaluation_costs_view,
         name="evaluation-costs",
     ),
-    path("evaluation/<uuid:evaluation_id>/end", views.end_page_view, name="end"),
+    path(
+        "evaluation/<uuid:evaluation_id>/policy-costs",
+        submission_views.evaluation_policy_costs_view,
+        name="policy-costs",
+    ),
+    path(
+        "evaluation/<uuid:evaluation_id>/publication-intention",
+        submission_views.evaluation_publication_intention_view,
+        name="publication-intention",
+    ),
+    path(
+        "evaluation/<uuid:evaluation_id>/documents",
+        submission_views.evaluation_documents_view,
+        name="documents",
+    ),
+    path("evaluation/<uuid:evaluation_id>/end", submission_views.end_page_view, name="end"),
 ]
 
 outcome_measure_urlpatterns = [
     path(
         "evaluation/<uuid:evaluation_id>/outcome-measures",
-        views.initial_outcome_measure_page_view,
+        submission_views.initial_outcome_measure_page_view,
         name="outcome-measures",
     ),
     path(
         "evaluation/<uuid:evaluation_id>/outcome-measures/last",
-        views.last_outcome_measure_page_view,
+        submission_views.last_outcome_measure_page_view,
         name="outcome-measure-last",
     ),
     path(
         "evaluation/<uuid:evaluation_id>/outcome-measures/first",
-        views.first_outcome_measure_page_view,
+        submission_views.first_outcome_measure_page_view,
         name="outcome-measure-first",
     ),
     path(
         "evaluation/<uuid:evaluation_id>/outcome-measures/last",
-        views.last_outcome_measure_page_view,
+        submission_views.last_outcome_measure_page_view,
         name="outcome-measure-last",
     ),
     path(
         "evaluation/<uuid:evaluation_id>/outcome-measures/<int:outcome_measure_id>",
-        views.outcome_measure_page_view,
+        submission_views.outcome_measure_page_view,
         name="outcome-measure-page",
     ),
     path(
         "evaluation/<uuid:evaluation_id>/outcome-measures/<int:outcome_measure_id>/delete",
-        views.delete_outcome_measure_page_view,
+        submission_views.delete_outcome_measure_page_view,
         name="outcome-measure-delete",
     ),
-    path("evaluation/<uuid:evaluation_id>/end", views.end_page_view, name="end"),
+    path("evaluation/<uuid:evaluation_id>/end", submission_views.end_page_view, name="end"),
 ]
 
 urlpatterns = urlpatterns + api_urlpatterns + evaluation_entry_urlpatterns + outcome_measure_urlpatterns
